armbian-config-Set-CPU-speed-and-governor-fix-and-warning
A change in lib/functions/rootfs/distro-agnostic.sh disables  cpufrequtils by default.

Armbian-config CPU - Set CPU speed and governor requires cpufrequtils to be enabled in order to work.

Cpufrequtils is deprecated and troublesome on some boards but is still used and wanted by forum members.

This change will add 'Deprecated' to the armbian-config menu item label and modify the associated job script to allow cpufrequtils to be enabled with a warning.